+++ mercadopago/includes/module/settings/TicketSettings.php
@@ -172,9 +172,8 @@ class TicketSettings extends AbstractSettings
 
         $payment_methods = $this->mercadopago->getPaymentMethods();
         foreach ($payment_methods as $payment_method) {
-            if ($payment_method['type'] != 'credit_card' &&
-                $payment_method['type'] != 'debit_card' &&
-                $payment_method['type'] != 'prepaid_card' &&
+            if ($payment_method['type'] == 'ticket' &&
+                Tools::strtolower($payment_method['id']) != 'meliplace' &&
                 !in_array($payment_method['id'], $this->getTicketExcludedMethods())
             ) {
                 $pm_id = $payment_method['id'];